+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Let users bundle arbitrary resources, have them extracted in cache, and used as `include` or `link` paths ([pull #43](https://github.com/bytedeco/javacpp/pull/43))
  * Fix potential formatting issues with `OutOfMemoryError` thrown from `Pointer`
  * Fix `Loader.getCallerClass()` ([pull #175](https://github.com/bytedeco/javacpp/pull/175))
  * Fix `Parser` not considering empty `class`, `struct`, or `union` declarations as opaque forward declarations